Restruct component layer and add stop run style in tool bar
Separate MainEditorController into components, and create a new root FXML named app-base with a controller named BaseController. Its child component is EditorController, which has two child components: EditorToolBar and EditorMenuBar. Additionally, change the Run button in EditorToolBar to a toggle button with a stop style, and write the stop logic in EditorToolBarController.